ai-service/requirements.txt
anthropic>=0.8.0
pypdf>=3.17.0
pdfplumber>=0.10.3
pydantic>=2.5.2
boto3>=1.34.14
python-dotenv>=1.0.0
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
pytest>=7.4.3
pytest-cov>=4.1.0
httpx>=0.25.2
psycopg2-binary>=2.9.9
